http://www.elvastower.com/forums/index.php?/topic/23644-signal-shields-wrong/ Possible fix for objects with alpha channels but without blending enabled appearing wrong since X2013.
git-svn-id: http://svn.uktrainsim.com/svn/openrails/trunk@2014 a9f3cfff-6c41-7446-b0ca-052bd761e6eb

+++ Source/RunActivity/Viewer3D/Shaders.cs
@@ -172,7 +172,7 @@ namespace ORTS
 
         public Texture2D OverlayTexture { set { overlayTexture.SetValue(value); } }
 
-        public byte ReferenceAlpha { set { referenceAlpha.SetValue(value / 255f); } }
+        public int ReferenceAlpha { set { referenceAlpha.SetValue(value / 255f); } }
 
         public SceneryShader(GraphicsDevice graphicsDevice)
             : base(graphicsDevice, "SceneryShader")
